Split evaluate.py by interface: seeds.py, cost.py, toml_utils writer; oracle to tests
evaluate.py held seven unrelated interfaces: the seed-pool registry, the
validation gate, the NN JSON writer, the cost function, a TOML serializer,
and three helpers no production path called. It now keeps the gate and the
two parameter writers.

- seeds.py: every *_SEED_OFFSET (the 7M sweep pool joins the registry
  instead of living in param_sweep.py), make_reserved_seeds,
  make_confirmatory_pools; test_seed_offsets.py asserts the full list and
  that no other package module defines an offset.
- cost.py: _softplus, dv_cost, compute_cost (numerics untouched).
- toml_utils.py: write_toml + set_dot_path.
- tests/fixtures/subprocess_oracle.py: the Rust-CLI oracle that backs
  test_pyo3_matches_subprocess (its only caller).
- Deleted with their tests: log_cap (deprecated reference) and
  patch_toml_mc_seed (test-only).

42 consumers rewired (package, experiments, paper scripts, tests); no
re-export aliases left behind. Stale "evaluate.py routes them" comments
in param_spaces.py now point at deploy_overrides.

Closes #71
