+++ articles/paper/scripts/objective_centering_eval.py
@@ -78,8 +78,8 @@ def extract_convergence(jsonl_path: str, n_pop: int, n_sims: int) -> list[list]:
 def _eval_one(label: str, run_dir: str, toml: str, n_sims_train: int, n_eval: int) -> dict:
     import aerocapture_rs
     from aerocapture.training.deploy_overrides import resolve_eval_toml
-    from aerocapture.training.evaluate import STRESS_EVAL_SEED_OFFSET, make_reserved_seeds
     from aerocapture.training.parquet_output import FINAL_COLUMNS, FINAL_RECORD_INDICES
+    from aerocapture.training.seeds import STRESS_EVAL_SEED_OFFSET, make_reserved_seeds
     from aerocapture.training.toml_utils import load_toml_with_bases
 
     scheme_dir = REPO / "training_output" / run_dir
