+++ articles/paper/scripts/robustness_retrain_eval.py
@@ -56,8 +56,8 @@ OUT = REPO / "articles/paper/data/robustness_retrain.json"
 def _eval_one(label: str, run_dir: str, toml: str, n_sims: int) -> dict:
     import aerocapture_rs
     from aerocapture.training.deploy_overrides import resolve_eval_toml
-    from aerocapture.training.evaluate import STRESS_EVAL_SEED_OFFSET, make_reserved_seeds
     from aerocapture.training.parquet_output import FINAL_COLUMNS, FINAL_RECORD_INDICES
+    from aerocapture.training.seeds import STRESS_EVAL_SEED_OFFSET, make_reserved_seeds
     from aerocapture.training.toml_utils import load_toml_with_bases
 
     scheme_dir = REPO / "training_output" / run_dir
