+++ tests/test_toml_patching.py
@@ -12,7 +12,7 @@ from pathlib import Path
 
 import pytest
 from aerocapture.training.encoding import decode_normalized
-from aerocapture.training.evaluate import patch_toml_mc_seed, write_guidance_toml
+from aerocapture.training.evaluate import write_guidance_toml
 from aerocapture.training.param_spaces import PARAM_SPACES
 
 from tests.fixtures.factories import make_normalized_individual
@@ -154,31 +154,6 @@ def test_toml_roundtrip_temp_file(scheme: str) -> None:
         assert "guidance" in parsed
     finally:
         written.unlink(missing_ok=True)
-
-
-class TestPatchTomlMcSeed:
-    def test_overrides_seed(self, tmp_path: Path) -> None:
-        base = tmp_path / "base.toml"
-        base.write_text('[monte_carlo]\nseed = 42\n\n[mission]\ntype = "aerocapture"\n')
-        patched = patch_toml_mc_seed(base, 99)
-        try:
-            with open(patched, "rb") as f:
-                data = tomllib.load(f)
-            assert data["monte_carlo"]["seed"] == 99
-            assert data["mission"]["type"] == "aerocapture"
-        finally:
-            patched.unlink(missing_ok=True)
-
-    def test_adds_seed_when_missing(self, tmp_path: Path) -> None:
-        base = tmp_path / "base.toml"
-        base.write_text('[mission]\ntype = "aerocapture"\n')
-        patched = patch_toml_mc_seed(base, 55)
-        try:
-            with open(patched, "rb") as f:
-                data = tomllib.load(f)
-            assert data["monte_carlo"]["seed"] == 55
-        finally:
-            patched.unlink(missing_ok=True)
 
 
 class TestWriteGuidanceTomlMcSeed:
